completed refactoring
