--- STEP_2_REVIEW_AND_APPROVAL.md
@@ -1,235 +1,0 @@
-# STEP 2: REVIEW & APPROVE - Refactoring Analysis
-
-**Date:** Current Session
-**Status:** All Step 1 analyses complete - Ready for approval
-**Files Analyzed:** 12 of 12
-
----
-
-## ✅ STEP 1 COMPLETE - Analysis Summary
-
-All 12 files have been analyzed following your systematic template. Below is the complete review for your approval.
-
----
-
-## 📊 Files Categorized by Action
-
-### ⚪ SKIP - Already Well-Architected (2 files)
-
-**1. useSpanLabeling.js (438 lines)** ✅
-- **Verdict:** Already has proper folder structure
-- **Structure:** useSpanLabeling/ with api/, config/, services/, utils/, __tests__/
-- **Action:** No refactoring needed - This IS the reference example
-
-**2. useVideoConceptState.js (291 lines)** ✅
-- **Verdict:** Already uses correct useReducer pattern
-- **Structure:** Single file with clear reducer, well-organized
-- **Action:** Optional organizational improvements only (low priority)
-
----
-
-### 🟡 NEEDS REFACTORING - Client Components (4 files)
-
-**3. WizardVideoBuilder.jsx (584 lines)** ⚠️
-- **Problems:**
-  - 9 separate useState calls (should be 1 useReducer)
-  - ~200 lines of business logic mixed in component
-  - ~100 lines of inline configuration
-- **Complexity:** MEDIUM
-- **Advantage:** 2 sibling components already refactored (StepCoreConcept/, StepQuickFill/)
-- **Proposed:** Extract to folder with hooks/, config/, utils/
-
-**4. useHighlightRendering.js (281 lines)** ⚠️
-- **Problems:**
-  - 186-line effect with mixed concerns
-  - Complex logic that should be pure functions
-  - Inline configuration and styles
-- **Complexity:** MEDIUM
-- **Proposed:** Extract to folder with utils/, config/, hooks/
-
-**5. PromptOptimizerContainer.jsx (716 lines)** ⚠️
-- **Problems:**
-  - Still too large despite previous refactoring
-  - Long event handlers with complex business logic
-  - 50+ values destructured from context
-- **Complexity:** MEDIUM-HIGH
-- **Proposed:** Extract business logic to custom hooks
-
-**6. StepAtmosphere.jsx (494 lines)** ⚠️
-- **Problems:**
-  - Flat file with ~392 lines of JSX with inline styles
-  - Should match StepCoreConcept/StepQuickFill pattern
-- **Complexity:** LOW-MEDIUM
-- **Advantage:** Can reuse hooks/components from sibling steps
-- **Proposed:** Extract to folder following sibling pattern
-
----
-
-### 🔴 NEEDS REFACTORING - Server Services (4 files)
-
-**7. EnhancementService.js (582 lines)** ⚠️
-- **Problems:**
-  - 377-line method (getEnhancementSuggestions)
-  - 87-line while loop for fallback regeneration
-  - Inline schema definitions
-- **Complexity:** HIGH
-- **Key Issue:** Fallback regeneration logic should be separate service
-- **Proposed:** Extract FallbackRegenerationService, SuggestionProcessor, config/
-
-**8. QualityFeedbackSystem.js (556 lines)** ⚠️
-- **Problems:**
-  - Single class with 4 distinct responsibilities
-  - ~140 lines feature extraction, ~105 lines quality assessment
-  - Hardcoded configuration throughout
-- **Complexity:** MEDIUM
-- **Proposed:** Split into FeatureExtractor, QualityAssessor, QualityModel, FeedbackRepository
-
-**9. VideoPromptService.js (563 lines)** ⚠️
-- **Problems:**
-  - 217-line method with 7 inline constraint builder functions
-  - ~250 lines of hardcoded configuration (45% of file!)
-  - 112-line method with complex category mapping
-- **Complexity:** MEDIUM-HIGH
-- **Key Issue:** File is essentially configuration disguised as code
-- **Proposed:** Extract all config, create specialized services
-
----
-
-### ❌ MISCLASSIFIED - Need Relocation (3 files)
-
-**10. SemanticCacheEnhancer.js (366 lines)** ❌
-- **Problem:** Complex service in utils/
-- **Action:** Move to services/cache/SemanticCacheService.js
-- **Complexity:** MEDIUM
-
-**11. ConcurrencyLimiter.js (337 lines)** ❌
-- **Problem:** Stateful service in utils/
-- **Action:** Move to services/concurrency/ConcurrencyService.js
-- **Complexity:** LOW-MEDIUM
-
-**12. validation.js (285 lines)** ⚠️
-- **Problem:** Schemas (config) mixed with validation logic
-- **Action:** Split schemas to config/, validation to middleware/
-- **Complexity:** LOW
-
----
-
-## 🎯 RECOMMENDED REFACTORING ORDER
-
-Based on impact, complexity, and dependencies:
-
-### Phase 1: Quick Wins (Low-Medium Complexity)
-1. **StepAtmosphere.jsx** - Clear pattern from siblings, straightforward
-2. **validation.js** - Simple file splitting
-3. **ConcurrencyLimiter.js** - Clean move to services/
-4. **SemanticCacheEnhancer.js** - Well-organized, easy extraction
-
-### Phase 2: Core Improvements (Medium Complexity)
-5. **WizardVideoBuilder.jsx** - Follow sibling pattern, complete wizard consistency
-6. **useHighlightRendering.js** - Extract pure functions and config
-7. **QualityFeedbackSystem.js** - Clear service boundaries
-8. **VideoPromptService.js** - Extract configuration
-
-### Phase 3: Complex Refactoring (High Complexity)
-9. **PromptOptimizerContainer.jsx** - Extract business logic to hooks
-10. **EnhancementService.js** - Extract fallback regeneration service
-
----
-
-## 📝 DETAILED ANALYSIS LOCATIONS
-
-All detailed Step 1 analyses are documented in:
-- `/Users/bryceharmon/Desktop/prompt-builder/REFACTORING_ANALYSIS_SUMMARY.md`
-
-Each analysis includes:
-- ✓ Exported functions/classes
-- ✓ Distinct responsibilities with line ranges
-- ✓ External dependencies
-- ✓ Hardcoded configuration
-- ✓ Architectural problems
-- ✓ Proposed refactoring structure
-- ✓ Estimated line counts
-- ✓ Migration complexity
-- ✓ Reference patterns
-
----
-
-## 🚦 APPROVAL OPTIONS
-
-Choose how to proceed with Step 3 (Execute Refactoring):
-
-### Option A: Start with Quick Wins (Recommended)
-Refactor files 1-4 from Phase 1 first:
-- StepAtmosphere.jsx
-- validation.js
-- ConcurrencyLimiter.js
-- SemanticCacheEnhancer.js
-
-**Advantage:** Build momentum with lower-risk refactorings
-
-### Option B: Prioritize by Business Impact
-Focus on high-traffic files:
-- WizardVideoBuilder.jsx (complete wizard consistency)
-- PromptOptimizerContainer.jsx (main feature component)
-- EnhancementService.js (core suggestion logic)
-
-**Advantage:** Immediate impact on most-used features
-
-### Option C: One File at a Time
-Choose any single file to start with.
-
-**Advantage:** Cautious approach, validate each step
-
-### Option D: Approve All, Execute in Order
-Approve entire plan, execute Phase 1 → Phase 2 → Phase 3.
-
-**Advantage:** Systematic completion of all refactoring
-
----
-
-## ⚠️ IMPORTANT NOTES
-
-### Line Count Guidelines (Not Hard Caps)
-Per your clarification, line counts are **guidelines for architectural quality**, not hard caps:
-- Focus is on architectural problems (multiple useState, mixed concerns, inline config)
-- Not on arbitrary line count limits
-- Some well-architected orchestrators may be 500+ lines (that's okay)
-
-### Files to Skip
-- **useSpanLabeling.js** - Already refactored properly ✅
-- **useVideoConceptState.js** - Already well-architected ✅
-
-### Migration Risks
-High complexity files need extra testing:
-- **EnhancementService.js** - 87-line while loop requires careful extraction
-- **PromptOptimizerContainer.jsx** - Many interdependent async operations
-
----
-
-## ✅ NEXT STEPS
-
-Please review and provide approval:
-
-1. **Approve entire plan** (proceed to Phase 1)
-2. **Select specific files** to start with
-3. **Request modifications** to any analysis
-4. **Ask questions** about any proposed changes
-
-Once approved, I'll proceed to **Step 3: Execute Refactoring** for the selected files.
-
----
-
-## 📋 Validation After Each Refactoring
-
-For each file refactored, I will:
-- ✓ Create backup file first
-- ✓ Preserve public API (no breaking changes)
-- ✓ Run tests to verify functionality
-- ✓ Check line counts (guidelines)
-- ✓ Create REFACTORING_SUMMARY.md
-- ✓ Verify imports still work
-
----
-
-**Ready for your approval to proceed! 🚀**
-
