+++ server/src/services/enhancement/services/StyleTransferService.js
@@ -1,0 +1,62 @@
+import { logger } from '../../../infrastructure/Logger.js';
+import { STYLE_DEFINITIONS, DEFAULT_STYLE } from '../config/styleDefinitions.js';
+
+/**
+ * Service responsible for transferring text between different writing styles
+ */
+export class StyleTransferService {
+  constructor(claudeClient) {
+    this.claudeClient = claudeClient;
+  }
+
+  /**
+   * Transfer text from one style to another
+   * @param {string} text - Text to transform
+   * @param {string} targetStyle - Target style (technical, creative, academic, casual, formal)
+   * @returns {Promise<string>} Transformed text
+   */
+  async transferStyle(text, targetStyle) {
+    const styleConfig = STYLE_DEFINITIONS[targetStyle] || STYLE_DEFINITIONS[DEFAULT_STYLE];
+
+    const styleTransferPrompt = this._buildStyleTransferPrompt(text, targetStyle, styleConfig);
+
+    try {
+      const response = await this.claudeClient.complete(styleTransferPrompt, {
+        maxTokens: 1024,
+        temperature: 0.7,
+      });
+
+      return response.content[0].text.trim();
+    } catch (error) {
+      logger.warn('Failed to transfer style', { error });
+      return text; // Return original on error
+    }
+  }
+
+  /**
+   * Build style transfer prompt
+   * @private
+   */
+  _buildStyleTransferPrompt(text, targetStyle, styleConfig) {
+    return `Transform the following text to ${targetStyle} style while preserving its core meaning and information.
+
+Original text: "${text}"
+
+Target style characteristics:
+- Formality level: ${styleConfig.formality}
+- Language type: ${styleConfig.jargon}
+- Structure: ${styleConfig.structure}
+- Tone: ${styleConfig.tone}
+- Examples style: ${styleConfig.examples}
+
+Requirements:
+1. Maintain all factual information from the original
+2. Adapt vocabulary to match the target style
+3. Restructure sentences appropriately for the style
+4. Preserve the core message and intent
+5. Make it feel natural in the new style
+
+Provide ONLY the transformed text, no explanations:`;
+  }
+}
+
